mm_server/requirements.txt
fastapi
llama_index
numpy
onnxruntime
opencv_python_headless
Pillow
pyclipper
pydantic
Requests
rich
Shapely
torch==2.2.0
uvicorn
setuptools==69.5.1
llama-index-embeddings-huggingface
llama-index-embeddings-clip
regex 
tqdm
cachetools
git+https://github.com/openai/CLIP.git
faster-whisper
